Refactor the code in the connections module into the credentials and environments modules
