+++ dbt/adapters/duckdb/__init__.py
@@ -1,6 +1,6 @@
 from dbt.adapters.base import AdapterPlugin
 from dbt.adapters.duckdb.connections import DuckDBConnectionManager  # noqa
-from dbt.adapters.duckdb.connections import DuckDBCredentials
+from dbt.adapters.duckdb.credentials import DuckDBCredentials
 from dbt.adapters.duckdb.impl import DuckDBAdapter
 from dbt.include import duckdb
